Refactored the multi reader.
