+++ swim_utilities/byte_channel/Cargo.toml
@@ -8,6 +8,7 @@ bytes = "1.0"
 tokio = { version = "1.1.1", features = ["io-util"] }
 parking_lot = "0.11.0"
 waker-fn = "1.1.0"
+slab = "0.4.5"
 
 [dev-dependencies]
 tokio = { version = "1.1.1", features = ["rt-multi-thread", "test-util", "macros"] }
@@ -15,7 +16,6 @@ futures = "0.3.4"
 futures-util = "0.3.4"
 criterion = { version = "0.3", features = ["async_tokio"] }
 rand = "0.7"
-slab = "0.4.5"
 
 [[bench]]
 name = "async_throughput"
